All util function in one file
