+++ src/utils.js
@@ -1,5 +1,13 @@
-import { getRandomInteger } from "./getRandomInteger";
-import { returnRandomItem } from "./returnRandomItem";
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const returnRandomItem = (array) => {
+  const randomIndex = getRandomInteger(0, array.length - 1);
+  return array[randomIndex];
+};
 
 const LOREM_SENTENCES = [
   `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
